requirements.txt
flet==0.6.0
cryptography==39.0.1
Pillow==9.3.0
opencv-python==4.6.0.66
rembg==2.0.29
reportlab==3.6.11
packaging==24.1
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.0_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=f49b82715dce6365f75eddcb0a9bb47d0d46feaf14bbc739dfcbd677b7073f5b
